Extract some widgets from pages for reutilization
