+++ lib/core/constants/app_dimensions.dart
@@ -1,0 +1,3 @@
+class AppDimensions {
+  static const padding =  24.0; 
+}
