+++ lib/core/constants/app_failures_messages.dart
@@ -1,7 +1,7 @@
 class AppFailureMessages {
-  static const String SERVER_FAILURE = 'Server Failure';
+  static const String SERVER_FAILURE = 'Aconteceu um problema com o servidor, tente novamente mais tarde.';
   static const String INVALID_INPUT_FAILURE =
-      'Invalid input - The name must not be empty';
-  static const String NO_USER_CACHED = 'No user cached';
-  static const String NO_INTERNET_CONNECTION = 'No internet connection';
+      'Entrada inválida, o nome de busca não deve ser vazio';
+  static const String NO_USER_CACHED = 'Nenhum usuário salvo';
+  static const String NO_INTERNET_CONNECTION = 'Sem internet. Verifique sua conexão e tente novamente.';
 }
